+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
  - Added `ToColorAttachment` trait implementation for `Texture2dMultisample`.
  - Added `Texture2dMultisample::as_surface` method.
+ - Changed `VertexFormat` to take a `Cow<'static, str>` instead of a `String`.
  - Fixed a stack overflow in release mode.